+++ tests/Controller/SubtractNamedParams2Test.php
@@ -9,7 +9,7 @@ use OV\JsonRPCAPIBundle\RPC\V1\Subtract2\Subtract2Request;
 use OV\JsonRPCAPIBundle\RPC\V1\Subtract2Method;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
-final class SubtractNamedParams2Test extends AbstractTest
+final class SubtractNamedParams2Test extends AbstractControllerTestCase
 {
     public function testRpcCallWithNamedParameters()
     {